+++ src/common/js/index.js
@@ -10,7 +10,6 @@ import { RunContext } from '../classes/run-context.class.js';
 import './global.js';
 import './constants.js';
 import './configure.js';
-import './trace.js';
 
 // provides base classes in AppPages global object
 AppPages.DisplaySet = DisplaySet;