+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Define new allowedInMenu() function, thus bumping minor candidate version number + enumerate() becomes async
     - configure() now warns for unmanaged keys
     - Introduce new AppPages.MenuItem class
+    - Use pwix:logger universal logger
 
 ### 1.2.0